+++ osu.Game/Beatmaps/Beatmap.cs
@@ -134,11 +134,11 @@ namespace osu.Game.Beatmaps
 
         public double DistanceSpacing { get; set; } = 1.0;
 
-        public int GridSize { get; set; }
+        public float GridSize { get; set; }
 
         public PositionSnapGridType GridType { get; set; }
 
-        public int GridRotation { get; set; }
+        public float GridRotation { get; set; }
 
         public Vector2 GridOffset { get; set; } = new Vector2(256, 192);
